get route for user
